Move src files to src
