+++ src/contexts/DetectionResultContext.js
