+++ src/screens/Camera/CameraPage.js
