+++ src/screens/Camera/CameraPagesNavigator.js
