+++ src/screens/Camera/ImgDetect/ImageDetectPage.js
