+++ src/screens/Camera/ImgDetect/ImageDetectResultList.js
