+++ src/screens/Camera/ImgDetect/ImageDetectStyles.js
